Rename and modify exampleThemeStorage.ts to fit chayns requirements
The file exampleThemeStorage.ts was renamed to chaynsStorage.ts to reflect its new orientation towards managing Chayns data as opposed to themes. The types were updated to match the new purpose of the file. The new type Chayns was created and used in the new ChaynsStorage type. The previous 'toggle' method, which was theme specific, was replaced with a 'setValue' method that accepts a Chayns instance. The method tied to the storage instantiation was also adjusted to handle Chayns data.