Refactor received chayns data handling process
Refactored the way we handle chayns data received from content scripts in the application. This includes safeguards like robust validation with 'zod' to ensure the received data follows the ChaynsStorageDataSchema. This will help in preventing unexpected behaviors due to invalid data. Further, the ColorLog utility was removed in favor of native console.log and console.info methods for better log readability. Also, changes were made in the project structure and file arrangement for better organization of the codebase.