Increase timeout delay in useChaynsEnvData.ts
The timeout delay for the function setupTimeout in useChaynsEnvData.ts has been increased from 5000 milliseconds to 7500 milliseconds. This change was made in order to provide a larger time window to process concurrent requests or operations and aims to reduce the chance of triggering a timeout when functions or operations run longer than expected.