Improve PersonFinder UI and add margin to TextInput.
The PersonFinderButton.tsx file was modified to improve the Person Finder modal's user interface. This incorporated wrapping the TextInput field within a Box. Also, a bottom margin was applied to the Box enclosing the TextInput to add a little more spacing in the modal dialog. This visual tweak enhances the legibility and improves the overall user experience in the modal.