Add call to content script for chayns data retrieval
In 'src/pages/popup/NoChaynsEnvFound.tsx', a call to the content script has been added. This is triggered each time the popup is opened, prompting the retrieval of chayns data from the current page. This helps to maintain updated chayns data for the user's active conversations, offering a seamless interaction experience.