Add avatar to popup component for more visual appeal
The popup component within 'src/pages/popup/Popup.tsx' has been modified to include an Avatar that will display the image associated with the data.siteId. This improves the visual appeal of the popup and allows for better user experience as they can visually identify elements. An Anchor element enclosing the Title has been moved within a Group with the Avatar.