Add NotLoggedInAlert component for unauthorized users
A new component 'NotLoggedInAlert' was added to the 'Popup.tsx'. It shows an informational alert message when the user is not authorized or logged in. The new alert component has a close button to hide the message. This change enhances the user experience by clearly informing the user about the login status on the website.