Add Github link and copyright text to Popup.tsx
Imported 'Button', 'Text' and 'IconBrandGithub' from '@mantine/core' and '@tabler/icons-react' to 'Popup.tsx'. These were used to add an additional section at the end of the popup that consists of a copyright note and a button that links to a Github page. This change provides users with easily accessible information about the source of the application, and adds a quick way for them to contribute or report issues.