Add data-testid to AccessToken Alert and create relevant tests
A 'data-testid' attribute has been added to the Alert component in AccessTokenStatus.tsx for better selection in testing. Newly created AccessTokenStatus.test.tsx contains tests to check the rendering of an alert message based on the state of the access token. This aids in ensuring the alert functions correctly under varied conditions.