Add unit tests for AccessToken availability check
This commit introduces unit tests for the function useIsAccessTokenAvailable, which checks the availability and validity of stored access tokens. Moreover, it replaces the direct usage of useTobitAccessTokenStorage hook and differenceInDays function in AccessTokenStatus component and its test file with useIsAccessTokenAvailable function. These changes aim to centralize the logic of access token availability checking and make the code base easier to maintain.